refactor(safa_utils): move to kernel
