+++ crates/kernel/Cargo.toml
@@ -5,9 +5,9 @@ version = "0.4.0"
 edition = "2024"
 
 [dependencies]
+safa-abi = { path = "../safa-abi" }
 noto-sans-mono-bitmap = { version = "0.3.0", features = ["size_20"] }
 
-safa-utils = { path = "../utils" }
 lazy_static = { version = "1.5.0", features = ["spin_no_std"] }
 spin = { version = "0.9.8", default-features = false, features = [
     "lazy",
@@ -23,7 +23,7 @@ limine = "0.5.0"
 hashbrown = { version = "0.15.2", features = ["nightly"] }
 
 serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde_json = { version = "1.0", git = "https://github.com/ObserverUnit/srede-json-frok", default-features = false, features = [
+serde_json = { version = "1.0", git = "https://github.com/safiworks/srede-json-frok", default-features = false, features = [
     "alloc",
 ] }
 
@@ -37,6 +37,7 @@ compiler_builtins = "0.1.158"
 hermit-dtb = "0.1.1"
 bitfield-struct = "0.11.0"
 rustc-demangle = "0.1.25"
+
 [dev-dependencies]
 
 [profile.dev]
