+++ crates/kernel/src/threading/cpu_context.rs
@@ -65,7 +65,7 @@ impl ContextStatus {
 }
 
 use alloc::{boxed::Box, sync::Arc};
-pub use safa_utils::abi::raw::processes::ContextPriority;
+pub use safa_abi::raw::processes::ContextPriority;
 
 #[derive(Debug, Clone)]
 /// A node representing a Thread in a thread queue
