+++ crates/kernel/src/utils/bstr.rs
