+++ crates/kernel/src/utils/display.rs
