+++ crates/kernel/src/utils/io.rs
