+++ crates/kernel/src/utils/mod.rs
@@ -1,10 +1,17 @@
 //! This mod is a wrapper around the [`safa_utils`] crate
 //! with a few additions
 
-pub use safa_utils::*;
 pub mod alloc;
+pub mod ansi;
+pub mod bstr;
+pub mod display;
 #[cfg(target_arch = "aarch64")]
 pub mod dtb;
+pub mod either;
 pub mod elf;
+pub mod io;
 pub mod locks;
+pub mod path;
+pub mod syscalls;
+pub mod types;
 pub mod ustar;
